docs(governance): RUNBOOK STEP 29P progress registry closeout v0
Registry-only closeout binding for merged capital risk sizing v1 offline slice.
Sets NEXT_RUNBOOK_STEP=RUNBOOK_STEP_29Q without starting canonical order intent
or runtime effects.
